requirements.txt
asgiref==3.3.4
boto3==1.17.62
botocore==1.20.62
bson==0.5.10
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
dataclasses==0.8
Django==2.2.24
djangorestframework==3.12.4
djongo==1.3.0
idna==2.10
jmespath==0.10.0
mysql-connector-python==8.0.24
Pillow==8.2.0
protobuf==3.15.8
pycparser==2.20
pydub==0.25.1
pymongo==3.11.4
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pysolr==3.9.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.15.0
sqlparse==0.2.4
typing-extensions==3.10.0.0
urllib3==1.26.4
watson-developer-cloud==0.26.1
